+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add `automaticallyHandleAnimations` option to toggle the default animation behavior.
 * Add `duration` and `curve` options in controller methods that start animations.
+* Add `thumbElevation` option to configure the size of the shadow around the thumb.
+* Add `elevation` option to progress bar indicators.
 * Update docs.
 * Update examples.